Fixed error with voxel value (changed to double and changed ints to floats).

+++ include/wp3_compressor/compressor.h
@@ -84,7 +84,7 @@ private:
   // Compression setup
   double octreeResolution;
   float normDist;
-  float voxValue;
+  double voxValue;
   Compressor pointCloudEncoder;
 
   // Box crop filter
